// packages/core/src/lib/proto-generated/roomdata.ts
export namespace com.iamteer.wcf {
// ...
    export namespace RoomData {
        export class RoomMember extends pb_1.Message {
            #one_of_decls: number[][] = [];
            constructor(data?: any[] | {
                wxid?: string;
                name?: string;
                state?: number;
            }) {
                super();
                pb_1.Message.initialize(this, Array.isArray(data) ? data : [], 0, -1, [], this.#one_of_decls);
                if (!Array.isArray(data) && typeof data == "object") {
                    if ("wxid" in data && data.wxid != undefined) {
                        this.wxid = data.wxid;
                    }
                    if ("name" in data && data.name != undefined) {
                        this.name = data.name;
                    }
                    if ("state" in data && data.state != undefined) {
                        this.state = data.state;
                    }
                }
            }
            get wxid() {
                return pb_1.Message.getFieldWithDefault(this, 1, "") as string;
            }
            set wxid(value: string) {
                pb_1.Message.setField(this, 1, value);
            }
            get name() {
                return pb_1.Message.getFieldWithDefault(this, 2, "") as string;
            }
            set name(value: string) {
                pb_1.Message.setField(this, 2, value);
            }
            get state() {
                return pb_1.Message.getFieldWithDefault(this, 3, 0) as number;
            }
            set state(value: number) {
                pb_1.Message.setField(this, 3, value);
            }
            static fromObject(data: {
                wxid?: string;
                name?: string;
                state?: number;
            }): RoomMember {
                const message = new RoomMember({});
                if (data.wxid != null) {
                    message.wxid = data.wxid;
                }
                if (data.name != null) {
                    message.name = data.name;
                }
                if (data.state != null) {
                    message.state = data.state;
                }
                return message;
            }
            toObject() {
                const data: {
                    wxid?: string;
                    name?: string;
                    state?: number;
                } = {};
                if (this.wxid != null) {
                    data.wxid = this.wxid;
                }
                if (this.name != null) {
                    data.name = this.name;
                }
                if (this.state != null) {
                    data.state = this.state;
                }
                return data;
            }
            serialize(): Uint8Array;
            serialize(w: pb_1.BinaryWriter): void;
            serialize(w?: pb_1.BinaryWriter): Uint8Array | void {
                const writer = w || new pb_1.BinaryWriter();
                if (this.wxid.length)
                    writer.writeString(1, this.wxid);
                if (this.name.length)
                    writer.writeString(2, this.name);
                if (this.state != 0)
                    writer.writeInt32(3, this.state);
                if (!w)
                    return writer.getResultBuffer();
            }
            static deserialize(bytes: Uint8Array | pb_1.BinaryReader): RoomMember {
                const reader = bytes instanceof pb_1.BinaryReader ? bytes : new pb_1.BinaryReader(bytes), message = new RoomMember();
                while (reader.nextField()) {
                    if (reader.isEndGroup())
                        break;
                    switch (reader.getFieldNumber()) {
                        case 1:
                            message.wxid = reader.readString();
                            break;
                        case 2:
                            message.name = reader.readString();
                            break;
                        case 3:
                            message.state = reader.readInt32();
                            break;
                        default: reader.skipField();
                    }
                }
                return message;
            }
            serializeBinary(): Uint8Array {
                return this.serialize();
            }
            static deserializeBinary(bytes: Uint8Array): RoomMember {
                return RoomMember.deserialize(bytes);
            }
        }
    }
}